+++ border-atari-env/examples/random_pong.rs
@@ -3,7 +3,7 @@ use border_atari_env::{
     BorderAtariAct, BorderAtariActRawFilter, BorderAtariEnv, BorderAtariEnvConfig, BorderAtariObs,
     BorderAtariObsRawFilter,
 };
-use border_core::{record::BufferedRecorder, util, Env as _, Policy};
+use border_core::{DefaultEvaluator, Env as _, Evaluator, Policy};
 
 type Obs = BorderAtariObs;
 type Act = BorderAtariAct;
@@ -43,21 +43,26 @@ fn main() -> Result<()> {
     env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info")).init();
     fastrand::seed(42);
 
-    // Creates Pong environment
-    let env_config = env_config("pong".to_string());
-    let mut env = Env::build(&env_config, 42)?;
-    env.open()?;
+    // Pong environment configuration
+    let env_config = env_config("pong".to_string()).render(true);
 
     // Creates a random policy
-    let n_acts = env.get_num_actions_atari();
-    let policy_config = RandomPolicyConfig {
-        n_acts: n_acts as _,
+    let mut policy = {
+        let policy_config = {
+            let n_acts = {
+                let env = Env::build(&env_config, 42)?;
+                env.get_num_actions_atari()
+            };
+            RandomPolicyConfig {
+                n_acts: n_acts as _,
+            }
+        };
+        RandomPolicy::build(policy_config)
     };
-    let mut policy = RandomPolicy::build(policy_config);
 
     // Runs evaluation
-    let mut recorder = BufferedRecorder::new();
-    let _ = util::eval_with_recorder(&mut env, &mut policy, 5, &mut recorder)?;
+    let env_config = env_config.render(true);
+    let _ = DefaultEvaluator::new(&env_config, 0, 5)?.evaluate(&mut policy);
 
     Ok(())
 }