+++ border-core/src/lib.rs
@@ -21,7 +21,7 @@
 //!
 //! # Policy
 //!
-//! [`Policy<E: Env>`] represents a policy, from which actions are sampled for 
+//! [`Policy<E: Env>`] represents a policy, from which actions are sampled for
 //! environment `E`. [`Policy::sample()`] takes `E::Obs` and emits `E::Act`.
 //! It could be probabilistic or deterministic.
 //!
@@ -32,7 +32,7 @@
 //! the agent's policy might be probabilistic for exploration, while in evaluation mode,
 //! the policy might be deterministic.
 //!
-//! [`Agent::opt()`] method does a single optimization step. The definition of an 
+//! [`Agent::opt()`] method does a single optimization step. The definition of an
 //! optimization step depends on each agent. It might be multiple stochastic gradient
 //! steps in an optimization step. Samples for training are taken from
 //! [`R: ReplayBufferBase`][`ReplayBufferBase`].
@@ -45,7 +45,7 @@
 //! [`ReplayBufferBase`] trait is an abstraction of replay buffers. For handling samples,
 //! there are two associated types: `PushedItem` and `Batch`. `PushedItem` is a type
 //! representing samples pushed to the buffer. These samples might be generated from
-//! [`Step<E: Env>`]. [`StepProcessorBase<E: Env>`] trait provides the interface 
+//! [`Step<E: Env>`]. [`StepProcessorBase<E: Env>`] trait provides the interface
 //! for converting [`Step<E: Env>`] into `PushedItem`.
 //!
 //! `Batch` is a type of samples taken from the buffer for training [`Agent`]s.
@@ -57,10 +57,10 @@
 //! [`SimpleReplayBuffer<O, A>`] implementats [`ReplayBufferBase`].
 //! This type has two parameters `O` and `A`, which are representation of
 //! observation and action in the replay buffer. `O` and `A` must implement
-//! [`SubBatch`], which has the functionality of storing samples, like `Vec<T>`, 
+//! [`SubBatch`], which has the functionality of storing samples, like `Vec<T>`,
 //! for observation and action. The associated types `PushedItem` and `Batch`
 //! are the same type, [`StdBatch`], representing sets of `(o_t, r_t, a_t, o_t+1)`.
-//! 
+//!
 //! [`SimpleStepProcessor<E, O, A>`] might be used with [`SimpleReplayBuffer<O, A>`].
 //! It converts `E::Obs` and `E::Act` into [`SubBatch`]s of respective types
 //! and generates [`StdBatch`]. The conversion process relies on trait bounds,
@@ -80,6 +80,7 @@
 //! [`SimpleStepProcessor`]: replay_buffer::SimpleStepProcessor
 //! [`SimpleStepProcessor<E, O, A>`]: replay_buffer::SimpleStepProcessor
 pub mod error;
+mod evaluator;
 pub mod record;
 pub mod replay_buffer;
 pub mod util;
@@ -91,4 +92,5 @@ pub use base::{
 };
 
 mod trainer;
+pub use evaluator::{Evaluator, DefaultEvaluator};
 pub use trainer::{SyncSampler, Trainer, TrainerConfig};